Update products
Update products pages and handling
