+++ lib/data/mock_data.dart
@@ -1,20 +1,3 @@
 import '../models/product.dart';
 
-List<Product> mockProducts = [
-  Product(
-    id: '1',
-    name: 'Yoga Mat',
-    description: 'High-quality, non-slip yoga mat.',
-    price: 299.99,
-    imageUrl: 'https://via.placeholder.com/150',
-    category: 'Fitness Equipment',
-  ),
-  Product(
-    id: '2',
-    name: 'Aromatherapy Oil',
-    description: 'Relaxing essential oils for aromatherapy.',
-    price: 129.99,
-    imageUrl: 'https://via.placeholder.com/150',
-    category: 'Wellness Products',
-  ),
-];
+List<Product> mockProducts = [];
